Crate miden_assembly_syntax

Re-exports

pub use self::sema::SemanticAnalysisError;

pub use miden_debug_types as debuginfo;

pub use miden_utils_diagnostics as diagnostics;

pub use semver;

Modules

arbitrary

This module exports helpers for implementing the Arbitrary trait on types that build on top of primitives provided by this crate.

ast

Abstract syntax tree (AST) components of Miden programs, modules, and procedures.

library

parser

prettier

sema

testing

Macros

assert_diagnostic

Assert that the given diagnostic/error value, when rendered to stdout, contains the given pattern

assert_diagnostic_lines

Like assert_diagnostic, but matches each non-empty line of the rendered output to a corresponding pattern.

parse_module

regex

Create a Pattern::Regex from the given input

source_file

Construct an ::alloc::sync::Arc<miden_core::debuginfo::SourceFile> from a string literal or expression, such that emitted diagnostics reference the file and line on which the source file was constructed.

Structs

DisplayHex

A display helper for formatting a slice of bytes as hex with different options using Rust’s builtin format language

Felt

A Felt backed by Plonky3’s Goldilocks field element.

ParseOptions

The set of options which can be used to control the behavior of the Parse trait.

Report

Core Diagnostic wrapper type.

Word

A unit of data consisting of 4 field elements.

Constants

MAX_REPEAT_COUNT

Maximum allowed iteration count for repeat.<count> blocks.

Traits

Parse

This trait is meant to be implemented by any type that can be parsed to a Module, to allow methods which expect a Module to accept things like:

PrimeCharacteristicRing

A commutative ring, R, with prime characteristic, p.

PrimeField64

A prime field ℤ/p with order, p < 2^64.