miden_assembly_syntax::ast

Module constants

Re-exports

pub use self::eval::ConstEnvironment;

pub use self::eval::ConstEvalError;

Modules

eval

Structs

Constant

Represents a constant definition in Miden Assembly syntax, i.e. const.FOO = 1 + 1.

Enums

ConstantExpr

Represents a constant expression or value in Miden Assembly syntax.

ConstantOp

Represents the set of binary arithmetic operators supported in Miden Assembly syntax.

ConstantValue

Represents a constant value in Miden Assembly syntax.

HashKind

Represents the type of the final value to which some string value should be converted.