miden_assembly_syntax::ast::constants::eval

Function expr

pub fn expr<Env>(
    value: &ConstantExpr,
    env: &mut Env,
) -> Result<ConstantExpr, <Env as ConstEnvironment>::Error>
where
    Env: ?Sized + ConstEnvironment,
    <Env as ConstEnvironment>::Error: From<ConstEvalError>,

Evaluate expr in env, producing a new ConstantExpr representing the value produced as the result of evaluation.

If expr could not be fully evaluated, e.g. due to external references which are not yet available, the returned expression may be only partially evaluated, or even entirely unevaluated.

It is up to env to determine how unresolved foreign symbols are to be handled. See the ConstEnvironment trait for more details.