miden_assembly_syntax::ast::path::arbitrary

Function constant_pathbuf_random_length

pub fn constant_pathbuf_random_length(min: u8) -> impl Strategy<Value = PathBuf>

Generate a PathBuf of random length, but at least min prefix components, that is valid for use with constant items.

The returned PathBuf will always have at least one component, the name of the constant.