miden_assembly_syntax::ast

Trait TypeResolver

pub trait TypeResolver<E> {
    // Required methods
    fn source_manager(&self) -> Arc<dyn SourceManager>;
    fn resolve_local_failed(&self, err: SymbolResolutionError) -> E;
    fn get_type(
        &mut self,
        context: SourceSpan,
        gid: GlobalItemIndex,
    ) -> Result<Type, E>;
    fn get_local_type(
        &mut self,
        context: SourceSpan,
        id: ItemIndex,
    ) -> Result<Option<Type>, E>;
    fn resolve_type_ref(
        &mut self,
        ty: Span<&Path>,
    ) -> Result<SymbolResolution, E>;

    // Provided method
    fn resolve(&mut self, ty: &TypeExpr) -> Result<Option<Type>, E> { ... }
}

Abstracts over resolving an item to a concrete [Type], using one of:

A GlobalItemIndex

An ItemIndex

A Path

A TypeExpr

Since type resolution happens in two different contexts during assembly, this abstraction allows us to share more of the resolution logic in both places.

NOTE: Most methods of this trait take a mutable reference to the resolver, so that the resolver can mutate its own state as necessary during resolution (e.g. to manage a cache, or other side table-like data structures).

Required Methods

fn source_manager(&self) -> Arc<dyn SourceManager>

fn resolve_local_failed(&self, err: SymbolResolutionError) -> E

Should be called by consumers of this resolver to convert a SymbolResolutionError to the error type used by the TypeResolver implementation.

fn get_type( &mut self, context: SourceSpan, gid: GlobalItemIndex, ) -> Result<Type, E>

Get the [Type] corresponding to the item given by gid

fn get_local_type( &mut self, context: SourceSpan, id: ItemIndex, ) -> Result<Option<Type>, E>

Get the [Type] corresponding to the item in the current module given by id

fn resolve_type_ref(&mut self, ty: Span<&Path>) -> Result<SymbolResolution, E>

Attempt to resolve a symbol path, given by a TypeExpr::Ref, to an item

Provided Methods

fn resolve(&mut self, ty: &TypeExpr) -> Result<Option<Type>, E>

Resolve a TypeExpr to a concrete [Type]