miden_assembly_syntax::ast

Type Alias ErrorMsg

pub type ErrorMsg = Immediate<Arc<str>>;

The type of error messages used in MASM assertions.

Aliased Type

pub enum ErrorMsg {
    Value(Span<Arc<str>>),
    Constant(Ident),
}

Variants

Value(Span<Arc<str>>)

A literal integer value, either decimal or hex-encoded

Constant(Ident)

A constant identifier

This must refer to a constant definition in the current module.

All immediates of this type are folded to Value during semantic analysis, once all constant definitions are evaluated.