miden_assembly_syntax

Module parser

Structs

Lexer

The lexer that is used to perform lexical analysis Miden Assembly grammar. The lexer implements the Iterator trait, so in order to retrieve the tokens, you simply have to iterate over it.

ModuleParser

This is a wrapper around the lower-level parser infrastructure which handles orchestrating all of the pieces needed to parse a ast::Module from source, and run semantic analysis on it.

Scanner

Scanner handles the low-level details of reading characters from a raw input stream of bytes. It decodes those bytes into UTF-8 characters, and associates each character with the [miden_debug_types::ByteIndex] at which it occurs.

WordValue

Enums

BinEncodedValue

Represents one of the various types of values that have a hex-encoded representation in Miden Assembly source files.

BinErrorKind

DocumentationType

Represents the scope of a given documentation comment

HexErrorKind

IntValue

Represents one of the various types of values that have a hex-encoded representation in Miden Assembly source files.

LiteralErrorKind

ParsingError

PushValue

Token

The token type produced by crate::parser::Lexer, and consumed by the parser.

Functions

parse_forms

This is used in tests to parse source as a set of raw ast::Forms rather than as a ast::Module.

read_modules_from_dir

Read the contents (modules) of this library from dir, returning any errors that occur while traversing the file system.