miden_assembly_syntax::parser

Function read_modules_from_dir

pub fn read_modules_from_dir(
    dir: impl AsRef<Path>,
    namespace: impl AsRef<Path>,
    source_manager: Arc<dyn SourceManager>,
    warnings_as_errors: bool,
) -> Result<impl Iterator<Item = Box<Module>>, Report>

Read the contents (modules) of this library from dir, returning any errors that occur while traversing the file system.

Errors may also be returned if traversal discovers issues with the modules, such as invalid names, etc.

Returns an iterator over all parsed modules.